chore: remove extra cast
fix the extra case, we're already dealing with uint64

+++ domain/sequence/state/state.go
@@ -75,7 +75,7 @@ func (s *State) ImportSequences(ctx context.Context, seqs map[string]uint64) err
 	for namespace, value := range seqs {
 		sequences = append(sequences, sequence{
 			Namespace: namespace,
-			Value:     uint64(value),
+			Value:     value,
 		})
 	}
